Homepage done all working

+++ egzaminas/routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CategoryController as C;
 use App\Http\Controllers\BookController as B;
+use App\Http\Controllers\HomeController as H;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,23 +23,24 @@ Route::get('/', function () {
 Auth::routes();
 
 Route::prefix('categories')->name('c_')->group(function () {
-    Route::get('/', [C::class, 'index'])->name('index');
-    Route::get('/create', [C::class, 'create'])->name('create');
-    Route::post('/create', [C::class, 'store'])->name('store');
-    Route::delete('/delete/{category}', [C::class, 'destroy'])->name('delete');
-    Route::get('/edit/{category}', [C::class, 'edit'])->name('edit');
-    Route::put('/edit/{category}', [C::class, 'update'])->name('update');
-    Route::get('/show/{category}', [C::class, 'show'])->name('show');
+    Route::get('/', [C::class, 'index'])->name('index')->middleware('gate:user');
+    Route::get('/create', [C::class, 'create'])->name('create')->middleware('gate:admin');
+    Route::post('/create', [C::class, 'store'])->name('store')->middleware('gate:admin');
+    Route::delete('/delete/{category}', [C::class, 'destroy'])->name('delete')->middleware('gate:admin');
+    Route::get('/edit/{category}', [C::class, 'edit'])->name('edit')->middleware('gate:admin');
+    Route::put('/edit/{category}', [C::class, 'update'])->name('update')->middleware('gate:admin');
+    Route::get('/show/{category}', [C::class, 'show'])->name('show')->middleware('gate:user');
 });
 
 Route::prefix('books')->name('b_')->group(function () {
-    Route::get('/', [B::class, 'index'])->name('index');
-    Route::get('/create', [B::class, 'create'])->name('create');
-    Route::post('/create', [B::class, 'store'])->name('store');
-    Route::delete('/delete/{book}', [B::class, 'destroy'])->name('delete');
-    Route::get('/edit/{book}', [B::class, 'edit'])->name('edit');
-    Route::put('/edit/{book}', [B::class, 'update'])->name('update');
-    Route::get('/show/{book}', [B::class, 'show'])->name('show');
+    Route::get('/', [B::class, 'index'])->name('index')->middleware('gate:user');
+    Route::get('/create', [B::class, 'create'])->name('create')->middleware('gate:admin');
+    Route::post('/create', [B::class, 'store'])->name('store')->middleware('gate:admin');
+    Route::delete('/delete/{book}', [B::class, 'destroy'])->name('delete')->middleware('gate:admin');
+    Route::get('/edit/{book}', [B::class, 'edit'])->name('edit')->middleware('gate:admin');
+    Route::put('/edit/{book}', [B::class, 'update'])->name('update')->middleware('gate:admin');
+    Route::get('/show/{book}', [B::class, 'show'])->name('show')->middleware('gate:user');
 });
 
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+//Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Route::get('/', [H::class, 'homeList'])->name('home')->middleware('gate:home');
